requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
pydantic==2.9.2
weasyprint==62.3
markdown-it-py==3.0.0
